+++ gneiss/plugin_setup.py
@@ -19,4 +19,5 @@ plugin = qiime2.plugin.Plugin(
     package='gneiss')
 
 importlib.import_module('gneiss.regression')
+importlib.import_module('gneiss.plot')
 importlib.import_module('gneiss.cluster')